### 2024-02-19 #629 @pachadotdev
Capybara
Submitting Author Name: Mauricio Vargas Sepulveda
Submitting Author Github Handle: @pachadotdev
Repository:  https://github.com/pachadotdev/capybara
Submission type: Pre-submission

---

```
Package: capybara
Title: Fast and Memory Efficient Fitting of Linear Models With High-Dimensional
    Fixed Effects
Version: 0.2
    person(
        given = "Mauricio",
        family = "Vargas Sepulveda",
        email = "m.sepulveda@mail.utoronto.ca",
    )
    Formula,
    generics,
    MASS,
    stats
    tidyr
Depends: R(>= 3.5.0)
Description: Fast and user-friendly estimation of generalized linear models with
    multiple fixed effects and cluster the standard errors. The method to obtain
    the estimated fixed-effects coefficients is based on Stammann (2018)
    <https://arxiv.org/abs/1707.01815> and Gaure (2013)
    <https://dx.doi.org/10.1016/j.csda.2013.03.024>.
License: Apache License (>= 2)
NeedsCompilation: yes
LinkingTo: cpp11, cpp11armadillo
    pachadotdev/cpp11armadillo
```

This helps to estimate linear models with many fixed effects.

People (mostly) in the social sciences that need multiple controls in their models. This is especially useful in Economics and International Relations.

Fixest, alpaca. This one has different design choices and a reduced number of dependencies.

Does not apply.

No.
