### 2024-03-03 6c9924b
minimal C++ optimizations
