+++ src/04_group_sums.cpp
@@ -13,13 +13,14 @@
 
   // Auxiliary variables (storage)
   Mat<double> num(P, 1);
+  std::vector<int> indexes;
 
   // Compute sum of weighted group sums
   double denom = 0.0;
 
   for (int j = 0; j < J; j++) {
     // Subset j-th group
-    integers indexes = as_cpp<integers>(jlist[j]);
+    indexes = as_cpp<std::vector<int>>(jlist[j]);
     int I = indexes.size();
 
     // Compute numerator of the weighted group sum
@@ -41,10 +42,9 @@
   return as_doubles_matrix(num);
 }
 
-[[cpp11::register]] doubles_matrix<>
-group_sums_spectral_(const doubles_matrix<> &M_r, const doubles_matrix<> &v_r,
-                     const doubles_matrix<> &w_r, const int K,
-                     const list &jlist) {
+[[cpp11::register]] doubles_matrix<> group_sums_spectral_(
+    const doubles_matrix<> &M_r, const doubles_matrix<> &v_r,
+    const doubles_matrix<> &w_r, const int K, const list &jlist) {
   // Types conversion
   Mat<double> M = as_Mat(M_r);
   Mat<double> v = as_Mat(v_r);
@@ -60,24 +60,33 @@ group_sums_spectral_(const doubles_matrix<> &M_r, const doubles_matrix<> &v_r,
   // Compute sum of weighted group sums
   double denom = 0.0;
 
+  // Precompute weights and values
+  std::vector<double> weights(J);
+  std::vector<std::vector<double>> values(J, std::vector<double>(P));
+
   for (int j = 0; j < J; j++) {
     // Subset j-th group
     integers indexes = as_cpp<integers>(jlist[j]);
     int I = indexes.size();
 
-    // Compute numerator of the weighted group sum given bandwidth 'L'
+    for (int i = 0; i < I; i++) {
+      weights[i] = w[indexes[i]];
+      for (int p = 0; p < P; p++) {
+        values[i][p] = M(indexes[i], p);
+      }
+    }
+
+    // Compute numerator of the weighted group sum
     num.zeros();
-    for (int p = 0; p < P; p++) {
-      for (int k = 1; k <= K; k++) {
-        for (int i = k; i < I; i++) {
-          num(p, 0) += M(indexes[i], p) * v(indexes[i - k]) * I / (I - k);
-        }
+    for (int p = 0; p < P; ++p) {
+      for (int i = 0; i < I; i++) {
+        num(p, 0) += values[i][p];
       }
     }
 
     // Compute denominator of the weighted group sum
     for (int i = 0; i < I; i++) {
-      denom += w(indexes[i]);
+      denom += weights[i];
     }
   }
 
@@ -86,8 +95,8 @@ group_sums_spectral_(const doubles_matrix<> &M_r, const doubles_matrix<> &v_r,
   return as_doubles_matrix(num);
 }
 
-[[cpp11::register]] doubles_matrix<>
-group_sums_var_(const doubles_matrix<> &M_r, const list &jlist) {
+[[cpp11::register]] doubles_matrix<> group_sums_var_(
+    const doubles_matrix<> &M_r, const list &jlist) {
   // Types conversion
   Mat<double> M = as_Mat(M_r);
 
@@ -99,6 +108,9 @@ group_sums_var_(const doubles_matrix<> &M_r, const list &jlist) {
   Mat<double> v(P, 1);
   Mat<double> V(P, P);
 
+  // Precompute values
+  std::vector<std::vector<double>> values(J, std::vector<double>(P));
+
   // Compute covariance matrix
   V.zeros();
   for (int j = 0; j < J; ++j) {
[…]
