### 2024-03-03 2634f5c
beta, eta and QR computation on C++ side
