+++ src/cpp11.cpp
@@ -6,64 +6,112 @@
 #include <R_ext/Visibility.h>
 
 // 01_center_variables.cpp
-doubles_matrix<> center_variables_(const doubles_matrix<> & V_r, const doubles & w_r, const list & klist, const double tol, const int maxiter);
-extern "C" SEXP _capybara_center_variables_(SEXP V_r, SEXP w_r, SEXP klist, SEXP tol, SEXP maxiter) {
+doubles_matrix<> center_variables_(const doubles_matrix<> & V_r, const doubles & v_sum_r, const doubles & w_r, const list & klist, const double tol, const int maxiter, bool sum_v);
+extern "C" SEXP _capybara_center_variables_(SEXP V_r, SEXP v_sum_r, SEXP w_r, SEXP klist, SEXP tol, SEXP maxiter, SEXP sum_v) {
   BEGIN_CPP11
-    return cpp11::as_sexp(center_variables_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(V_r), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(w_r), cpp11::as_cpp<cpp11::decay_t<const list &>>(klist), cpp11::as_cpp<cpp11::decay_t<const double>>(tol), cpp11::as_cpp<cpp11::decay_t<const int>>(maxiter)));
+    return cpp11::as_sexp(center_variables_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(V_r), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(v_sum_r), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(w_r), cpp11::as_cpp<cpp11::decay_t<const list &>>(klist), cpp11::as_cpp<cpp11::decay_t<const double>>(tol), cpp11::as_cpp<cpp11::decay_t<const int>>(maxiter), cpp11::as_cpp<cpp11::decay_t<bool>>(sum_v)));
   END_CPP11
 }
-// 02_get_alpha.cpp
+// 02_solve_beta.cpp
+doubles solve_beta_(const doubles_matrix<>& mx, const doubles_matrix<>& mnu, const doubles wtilde, double epsilon, bool weighted);
+extern "C" SEXP _capybara_solve_beta_(SEXP mx, SEXP mnu, SEXP wtilde, SEXP epsilon, SEXP weighted) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(solve_beta_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<>&>>(mx), cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<>&>>(mnu), cpp11::as_cpp<cpp11::decay_t<const doubles>>(wtilde), cpp11::as_cpp<cpp11::decay_t<double>>(epsilon), cpp11::as_cpp<cpp11::decay_t<bool>>(weighted)));
+  END_CPP11
+}
+// 03_get_alpha.cpp
 list get_alpha_(const doubles_matrix<> & p_r, const list & klist, const double tol);
 extern "C" SEXP _capybara_get_alpha_(SEXP p_r, SEXP klist, SEXP tol) {
   BEGIN_CPP11
     return cpp11::as_sexp(get_alpha_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(p_r), cpp11::as_cpp<cpp11::decay_t<const list &>>(klist), cpp11::as_cpp<cpp11::decay_t<const double>>(tol)));
   END_CPP11
 }
-// 03_group_sums.cpp
+// 03_solve_eta.cpp
[…]
