+++ src/cpp11.cpp
@@ -13,10 +13,10 @@ extern "C" SEXP _capybara_center_variables_(SEXP V_r, SEXP v_sum_r, SEXP w_r, SE
   END_CPP11
 }
 // 02_solve_beta.cpp
-doubles solve_beta_(const doubles_matrix<>& mx, const doubles_matrix<>& mnu, const doubles wtilde, double epsilon, bool weighted);
+doubles solve_beta_(const doubles_matrix<> & mx, const doubles_matrix<> & mnu, const doubles wtilde, double epsilon, bool weighted);
 extern "C" SEXP _capybara_solve_beta_(SEXP mx, SEXP mnu, SEXP wtilde, SEXP epsilon, SEXP weighted) {
   BEGIN_CPP11
-    return cpp11::as_sexp(solve_beta_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<>&>>(mx), cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<>&>>(mnu), cpp11::as_cpp<cpp11::decay_t<const doubles>>(wtilde), cpp11::as_cpp<cpp11::decay_t<double>>(epsilon), cpp11::as_cpp<cpp11::decay_t<bool>>(weighted)));
+    return cpp11::as_sexp(solve_beta_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(mx), cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(mnu), cpp11::as_cpp<cpp11::decay_t<const doubles>>(wtilde), cpp11::as_cpp<cpp11::decay_t<double>>(epsilon), cpp11::as_cpp<cpp11::decay_t<bool>>(weighted)));
   END_CPP11
 }
 // 03_get_alpha.cpp
@@ -27,31 +27,24 @@ extern "C" SEXP _capybara_get_alpha_(SEXP p_r, SEXP klist, SEXP tol) {
   END_CPP11
 }
 // 03_solve_eta.cpp
-doubles solve_eta_(const doubles_matrix<>& mx, const doubles_matrix<>& mnu, const doubles& nu, const doubles& beta);
+doubles solve_eta_(const doubles_matrix<> & mx, const doubles_matrix<> & mnu, const doubles & nu, const doubles & beta);
 extern "C" SEXP _capybara_solve_eta_(SEXP mx, SEXP mnu, SEXP nu, SEXP beta) {
   BEGIN_CPP11
-    return cpp11::as_sexp(solve_eta_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<>&>>(mx), cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<>&>>(mnu), cpp11::as_cpp<cpp11::decay_t<const doubles&>>(nu), cpp11::as_cpp<cpp11::decay_t<const doubles&>>(beta)));
+    return cpp11::as_sexp(solve_eta_(cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(mx), cpp11::as_cpp<cpp11::decay_t<const doubles_matrix<> &>>(mnu), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(nu), cpp11::as_cpp<cpp11::decay_t<const doubles &>>(beta)));
   END_CPP11
 }
 // 03_solve_eta.cpp
-doubles solve_eta2_(const doubles& yadj, const doubles_matrix<>& myadj, const doubles& offset, const doubles& eta);
+doubles solve_eta2_(const doubles & yadj, const doubles_matrix<> & myadj, const doubles & offset, const doubles & eta);
[…]
