+++ NEWS.md
@@ -1,3 +1,7 @@
+# capybara 0.3.5
+
+* Uses Mat<T> consistently for all matrix operations (avoids vectors)
+
 # capybara 0.3
 
 * Reduces memory footprint ~45% by moving some computation to Armadillo's side
