+++ dev/ropensci.R
@@ -1,0 +1,2 @@
+remotes::install_github("ropensci-review-tools/srr")
+srr::srr_stats_roxygen(category = c("regression"))
