### 2024-03-06 #629 @pachadotdev
> @pachadotdev There'll always be checks that do not apply to you package. 60/116 is enough to comply, so you could already call `check srr` yourself to confirm. Before doing that, however, could you please indicate how many more you think you may be able to comply with? We intentionally avoid hard limits other than minimal 50%, but it would be better if you could comply with, say, around 60% or more. Do you think that might be possible?

yes, if I move the scripts in dev to formal tests, it would be 70% i would say
