### 2024-03-17 8351409
back to dplyr, idk other way to fix negbin estimation
