+++ R/fenegbin.R
@@ -57,7 +57,6 @@ fenegbin <- function(
   k <- length(k.vars)
 
   # Generate temporary variable ----
-  setkeyv(data, k.vars)
   tmp.var <- temp_var_(data)
 
   # Drop observations that do not contribute to the log likelihood ----
@@ -92,7 +91,7 @@ fenegbin <- function(
   start_guesses_(beta.start, eta.start, y, X, beta, nt, wt, p, family)
 
   # Get names and number of levels in each fixed effects category ----
-  nms.fe <- lapply(data[, k.vars, with = FALSE], levels)
+  nms.fe <- lapply(select(data, all_of(k.vars)), levels)
   lvls.k <- vapply(nms.fe, length, integer(1))
 
   # Generate auxiliary list of indexes for different sub panels ----
