+++ R/helpers.R
@@ -47,7 +47,7 @@ partial_mu_eta_ <- function(eta, family, order) {
 
 temp_var_ <- function(data) {
   repeat {
-    tmp.var <- paste0(sample(letters, 5L, replace = TRUE), collapse = "")
+    tmp.var <- paste0("capybara_internal_variable_", sample(letters, 5L, replace = TRUE), collapse = "")
     if (!(tmp.var %in% colnames(data))) {
       break
     }
@@ -108,8 +108,7 @@ update_formula_ <- function(formula) {
 }
 
 model_frame_ <- function(data, formula, weights) {
-  setDT(data)
-  data <- data[, c(all.vars(formula), weights), with = FALSE]
+  data <- select(data, all_of(c(all.vars(formula), weights)))
 
   lhs <- names(data)[[1L]]
 
@@ -129,33 +128,38 @@ model_frame_ <- function(data, formula, weights) {
 check_response_ <- function(data, lhs, family) {
   if (family[["family"]] == "binomial") {
     # Check if 'y' is numeric
-    if (data[, is.numeric(get(lhs))]) {
+    if (is.numeric(pull(select(data, !!sym(lhs))))) {
       # Check if 'y' is in [0, 1]
-      if (data[, any(get(lhs) < 0.0 | get(lhs) > 1.0)]) {
+      if (nrow(filter(data, !!sym(lhs) < 0.0 | !!sym(lhs) > 1.0)) > 0L) {
         stop("Model response has to be within the unit interval.",
           call. = FALSE
         )
       }
     } else {
       # Check if 'y' is factor and transform otherwise
-      data[, (1L) := check_factor_(get(lhs))]
+      data <- mutate(data, !!sym(lhs) := check_factor_(!!sym(lhs)))
 
       # Check if the number of levels equals two
-      if (data[, length(levels(get(lhs)))] != 2L) {
+      if (nrow(summarise(data, n_levels = nlevels(!!sym(lhs)))) != 2L) {
         stop("Model response has to be binary.", call. = FALSE)
       }
 
       # Ensure 'y' is 0-1 encoded
-      data[, (1L) := as.numeric(get(lhs)) - 1.0]
+      ## if lhs is not numeric, convert it
+      if (!is.numeric(pull(select(data, !!sym(lhs))))) {
+        data <- mutate(data, !!sym(lhs) := as.numeric(!!sym(lhs)) - 1.0)
+      } else {
+        data <- mutate(data, !!sym(lhs) := !!sym(lhs) - 1.0)
+      }
     }
   } else if (family[["family"]] %in% c("Gamma", "inverse.gaussian")) {
     # Check if 'y' is strictly positive
-    if (data[, any(get(lhs) <= 0.0)]) {
+    if (nrow(filter(data, !!sym(lhs) <= 0.0)) > 0L) {
       stop("Model response has to be strictly positive.", call. = FALSE)
     }
   } else if (family[["family"]] != "gaussian") {
     # Check if 'y' is positive
-    if (data[, any(get(lhs) < 0.0)]) {
+    if (nrow(filter(data, !!sym(lhs) < 0.0)) > 0L) {
       stop("Model response has to be positive.", call. = FALSE)
     }
   }
@@ -168,13 +172,16 @@ drop_by_link_type_ <- function(data, lhs, family, tmp.var, k.vars, control) {
         # Drop observations that do not contribute to the log-likelihood
         ncheck <- nrow(data)
         for (j in k.vars) {
-          data[, (tmp.var) := mean(get(lhs)), by = eval(j)]
+          data <- data %>%
+            group_by(!!sym(j)) %>%
[…]
