+++ NAMESPACE
@@ -27,6 +27,7 @@ S3method(tidy,felm)
 S3method(vcov,apes)
 S3method(vcov,feglm)
 S3method(vcov,felm)
+export("%>%")
 export(apes)
 export(augment)
 export(bias_corr)
@@ -41,13 +42,22 @@ export(tidy)
 importFrom(Formula,Formula)
 importFrom(MASS,negative.binomial)
 importFrom(MASS,theta.ml)
-importFrom(data.table,":=")
-importFrom(data.table,.SD)
-importFrom(data.table,setDT)
-importFrom(data.table,setkeyv)
+importFrom(dplyr,across)
+importFrom(dplyr,all_of)
+importFrom(dplyr,filter)
+importFrom(dplyr,group_by)
+importFrom(dplyr,mutate)
+importFrom(dplyr,pull)
+importFrom(dplyr,select)
+importFrom(dplyr,summarise)
+importFrom(dplyr,ungroup)
+importFrom(dplyr,vars)
 importFrom(generics,augment)
 importFrom(generics,glance)
 importFrom(generics,tidy)
+importFrom(magrittr,"%>%")
+importFrom(rlang,":=")
+importFrom(rlang,sym)
 importFrom(stats,as.formula)
 importFrom(stats,binomial)
 importFrom(stats,gaussian)
