+++ DESCRIPTION
@@ -12,10 +12,12 @@ Authors@R: c(
         comment = c(ORCID = "0000-0003-1017-7574"))
     )
 Imports:
-    data.table,
+    dplyr,
     Formula,
     generics,
+    magrittr,
     MASS,
+    rlang,
     stats
 Suggests: 
     knitr,
