### 2024-03-17 bc9d596
slightly bigger memory usage with dplyr
