### 2024-03-19 #629 @ldecicco-USGS
Thank you @pachadotdev for this pre-submission inquiry. This package is within our statistical scope and we'd be happy to review a full submission. You'll need to open a new issue for the full submission.

(see issue 615 for a recent example of a submission, where the pre-submission is 614)
